Autonomous Tweaks before unbag time

+++ Command-Based-Test/src/org/usfirst/frc/team1099/robot/subsystems/Drivetrain.java
@@ -74,11 +74,11 @@ public class Drivetrain extends Subsystem {
     	updateDashboard();
     }
     
-    public double getLeftEncoder() {
+    public int getLeftEncoder() {
     	return leftDrive.getSelectedSensorPosition(0);
     }
     
-    public double getRightEncoder() {
+    public int getRightEncoder() {
     	return rightDrive.getSelectedSensorPosition(0);
     }
